+++ src/main/java/com/li64/tide/datagen/fabric/providers/assets/TideModelProvider.java
@@ -72,6 +72,7 @@ public class TideModelProvider extends FabricModelProvider {
         generator.generateFlatItem(TideItems.LUNAR_CALENDAR, ModelTemplates.FLAT_ITEM);
         generator.generateFlatItem(TideItems.CLIMATE_GAUGE, ModelTemplates.FLAT_ITEM);
         generator.generateFlatItem(TideItems.DEPTH_METER, ModelTemplates.FLAT_ITEM);
+        generator.generateFlatItem(TideItems.WEATHER_RADIO, ModelTemplates.FLAT_ITEM);
 
         generator.generateFlatItem(TideItems.ENCHANTED_POCKET_WATCH, ModelTemplates.FLAT_ITEM);
         generator.generateFlatItem(TideItems.DRAGONFIN_BOOTS, ModelTemplates.FLAT_ITEM);