+++ src/main/java/com/li64/tide/registries/TideItems.java
@@ -94,6 +94,7 @@ public class TideItems {
     public static final Item LUNAR_CALENDAR = register("lunar_calendar", LunarCalendarItem::new);
     public static final Item CLIMATE_GAUGE = register("climate_gauge", ClimateGaugeItem::new);
     public static final Item DEPTH_METER = register("depth_meter", DepthMeterItem::new);
+    public static final Item WEATHER_RADIO = register("weather_radio", WeatherRadioItem::new);
 
     public static final Item ENCHANTED_POCKET_WATCH = register("enchanted_pocket_watch", EnchantedPocketWatchItem::new, new Item.Properties()
             /*? if >=1.21*/.component(DataComponents.ENCHANTMENT_GLINT_OVERRIDE, true)
@@ -156,7 +157,7 @@ public class TideItems {
 
                     BAIT, LUCKY_BAIT, MAGNETIC_BAIT,
 
-                    POCKET_WATCH, LUNAR_CALENDAR, CLIMATE_GAUGE, DEPTH_METER,
+                    POCKET_WATCH, LUNAR_CALENDAR, CLIMATE_GAUGE, DEPTH_METER, WEATHER_RADIO,
 
                     ENCHANTED_POCKET_WATCH, STARLIGHT_BOW, MIDAS_FISHING_ROD, DRAGONFIN_BOOTS,